Refactor packages
